Move ServerConfig to separate file
